requirements.txt
discord.py
requests
Pillow
pydantic
python-dotenv
google-genai
beautifulsoup4